refactor(PermisionLauncher): object 클래스의 함수들을 top level에 선언한 후 클래스 이름 변경
